UDD - Score
결과창에 펫 데미지 추가

+++ Assets/Scripts/UI/Result/ResultSetter.cs
@@ -23,7 +23,7 @@ public class ResultSetter : MonoBehaviour
         }
 
         weaponName.text = name;
-        weaponDamage.text = damage.ToString();
+        weaponDamage.text = ((int)damage).ToString();
 
         graph.GetComponent<RectTransform>().sizeDelta = new Vector2(damage / ScoreManager.instance.GetTotalDamages() * 160, 24);
     }
